+++ src/components/FormElements/FormElements.tsx
@@ -1,7 +1,7 @@
 import Slider from 'rc-slider';
 import React, { ChangeEvent, useMemo } from 'react';
 import { css, cx } from '@emotion/css';
-import { dateTime, DateTime } from '@grafana/data';
+import { dateTime, DateTime, InterpolateFunction } from '@grafana/data';
 import {
   CodeEditor,
   DateTimePicker,
@@ -26,7 +26,7 @@ import {
 } from '../../constants';
 import { Styles } from '../../styles';
 import { LayoutSection, LocalFormElement, PanelOptions } from '../../types';
-import { ApplyWidth, FormatNumberValue, ToNumberValue } from '../../utils';
+import { ApplyWidth, FormatNumberValue, GetLayoutUniqueId, ToNumberValue } from '../../utils';
 
 /**
  * Properties
@@ -60,12 +60,24 @@ interface Props {
    * Section
    */
   section: LayoutSection | null;
+
+  /**
+   * Template variables interpolation function
+   */
+  replaceVariables: InterpolateFunction;
 }
 
 /**
  * Form Elements
  */
-export const FormElements: React.FC<Props> = ({ options, elements, onChangeElement, section, initial }) => {
+export const FormElements: React.FC<Props> = ({
+  options,
+  elements,
+  onChangeElement,
+  section,
+  initial,
+  replaceVariables,
+}) => {
   /**
    * Theme and Styles
    */
@@ -97,9 +109,9 @@ export const FormElements: React.FC<Props> = ({ options, elements, onChangeEleme
    */
   const visibleElements = useMemo(() => {
     return elements.filter((element) => {
-      return element.helpers.showIf({ elements });
+      return element.helpers.showIf({ elements, replaceVariables });
     });
-  }, [elements]);
+  }, [elements, replaceVariables]);
 
   return (
     <div data-testid={TestIds.formElements.root}>
@@ -107,7 +119,7 @@ export const FormElements: React.FC<Props> = ({ options, elements, onChangeEleme
         /**
          * Skip Hidden Elements
          */
-        if (section && element.section !== section.name) {
+        if (section && element.section !== GetLayoutUniqueId(section)) {
           return;
         }