+++ src/components/FormElementsEditor/FormElementsEditor.test.tsx
@@ -1,5 +1,6 @@
 import React from 'react';
 import { DragDropContext, DropResult } from 'react-beautiful-dnd';
+import { toDataFrame } from '@grafana/data';
 import { act, fireEvent, render, screen, waitFor, within } from '@testing-library/react';
 import {
   CodeEditorHeight,
@@ -456,6 +457,76 @@ describe('Form Elements Editor', () => {
 
     expect(elementSelectors.fieldType()).toBeInTheDocument();
     expect(elementSelectors.fieldAccept()).toBeInTheDocument();
+  });
+
+  /*
+   * Field Name Disabled
+   */
+  it('Should not allow to select field name if initial method is not Datasource', () => {
+    const elements = [
+      {
+        ...FormElementDefault,
+        id: 'element',
+        type: FormElementType.STRING,
+      },
+    ];
+
+    render(
+      getComponent({
+        value: elements,
+        onChange,
+        context: {
+          options: {
+            initial: {
+              method: RequestMethod.QUERY,
+            },
+          },
+        },
+      })
+    );
+    expect(selectors.root()).toBeInTheDocument();
+
+    /**
+     * Make Element is opened
+     */
+    const elementSelectors = openElement('element', FormElementType.STRING);
+
+    expect(elementSelectors.fieldNamePicker(true)).not.toBeInTheDocument();
+  });
+
+  /**
+   * Query Field Name Disabled
+   */
+  it('Should not allow to select query field name if initial method is not Query', () => {
+    const elements = [
+      {
+        ...FormElementDefault,
+        id: 'element',
+        type: FormElementType.STRING,
+      },
+    ];
+
+    render(
+      getComponent({
+        value: elements,
+        onChange,
+        context: {
+          options: {
+            initial: {
+              method: RequestMethod.DATASOURCE,
+            },
+          },
+        },
+      })
+    );
+    expect(selectors.root()).toBeInTheDocument();
+
+    /**
+     * Make Element is opened
+     */
+    const elementSelectors = openElement('element', FormElementType.STRING);
+
+    expect(elementSelectors.fieldFromQueryPicker(true)).not.toBeInTheDocument();
   });
 
   /**
@@ -1204,6 +1275,53 @@ describe('Form Elements Editor', () => {
       await act(() => fireEvent.change(elementSelectors.fieldAccept(), { target: { value: '.png' } }));
 
       expect(elementSelectors.fieldAccept()).toHaveValue('.png');
+    });
+
+    it('Should update query field name', async () => {
+      const elements = [{ ...FormElementDefault, id: 'id' }];
+      const context = {
+        options: {
+          initial: {
+            method: RequestMethod.QUERY,
+          },
+        },
+        data: [
+          toDataFrame({
+            fields: [
+              {
+                name: 'field1',
+                values: [],
+              },
+              {
+                name: 'field2',
+                values: [],
+              },
+            ],
+          }),
+        ],
+      };
+
+      render(getComponent({ value: elements, onChange, context }));
+
+      /**
+       * Open id element
+       */
+
+      const elementSelectors = openElement('id', FormElementDefault.type);
+
+      /**
+       * Change query field name
+       */
+      await act(() => fireEvent.change(elementSelectors.fieldFromQueryPicker(), { target: { value: 'field1' } }));
+
+      expect(elementSelectors.fieldFromQueryPicker()).toHaveValue('field1');
+
+      /**
+       * Clear query field name
+       */
+      await act(() => fireEvent.change(elementSelectors.fieldFromQueryPicker(), { target: { value: '' } }));
+
+      expect(elementSelectors.fieldFromQueryPicker()).toHaveValue('');
     });
 
     it('Should update showIf', async () => {