+++ src/types/form-element.ts
@@ -1,5 +1,9 @@
-import { SelectableValue } from '@grafana/data';
+import { InterpolateFunction, SelectableValue } from '@grafana/data';
 import { CodeLanguage, FormElementType } from '../constants';
+
+export type QueryField = SelectableValue<string> & {
+  refId?: string;
+};
 
 /**
  * Form Element Base
@@ -86,6 +90,11 @@ export interface FormElementBase {
    * Field Name
    */
   fieldName?: string;
+
+  /**
+   * Query Field
+   */
+  queryField?: QueryField;
 }
 
 /**
@@ -276,7 +285,10 @@ export type FormElement = FormElementBase &
 /**
  * Show If Helper
  */
-export type ShowIfHelper = (params: { elements: FormElement[] }) => boolean | undefined;
+export type ShowIfHelper = (params: {
+  elements: FormElement[];
+  replaceVariables: InterpolateFunction;
+}) => boolean | undefined;
 
 /**
  * Local Form Element